Add a new class to manage the more complex bit packing of the configuration register. Update everything for the new 17-bit data length (makes a lot of things less clean).
Still thinking about a way to combo the event counter read with
the more efficient 16 bit data length read that is nice.

+++ minervadaq/tests/TestSuite.h
@@ -13,6 +13,7 @@
 #include "log4cppHeaders.h"
 
 #include "Controller.h"
+#include "EChannelsConfigRegParser.h"
 #include "ECROC.h"
 #include "CRIM.h"
 #include "FrontEndBoard.h"
@@ -45,7 +46,7 @@ void FEBFPGAWriteReadTest( EChannels* channel, unsigned int nFEBs );
 // They should be called as a set as the sixth & seventh tests.
 void SetupChargeInjection( EChannels* channel, unsigned int nFEBs );
 unsigned char * ReadDPMTestData( ECROC * ecroc, unsigned int channel, 
-    unsigned int nFEBs, unsigned short pointer );
+    unsigned int nFEBs, unsigned int pointer );
 unsigned short int ReadDPMTestPointer( ECROC * ecroc, unsigned int channel, 
     unsigned int nFEBs );
